Add files via upload
Library management system is a project which aims in developing a computerized system to 
maintain all the daily work of library. The system would provide basic set of features to 
add/update members, add/update books, and manage check in specifications for the systems 
based on the client’s statement of need. The Library Management System is an application 
for assisting a librarian in managing a book library in a university. This project has many 
features which are generally not available in normal library management systems . It also has 
a facility of admin login through which the admin can monitor the whole system. The 
librarian after logging into his account i.e admin account can generate various reports such as 
issue report and book report. Library management system is a typical management 
Information system (MIS), its Development include the establishment and maintenance of 
back-end database and front-end application development aspects.